refactor: split onboard api key selection
