+++ crates/daemon/src/onboard_api_key_selection.rs
@@ -1,0 +1,110 @@
+use super::*;
+
+pub(super) fn resolve_api_key_env_selection(
+    options: &OnboardCommandOptions,
+    config: &mvp::config::LoongConfig,
+    default_api_key_env: String,
+    guided_prompt_path: GuidedPromptPath,
+    ui: &mut impl OnboardUi,
+    context: &OnboardRuntimeContext,
+) -> CliResult<String> {
+    let explicit_selection = if let Some(api_key_env) = options.api_key_env.as_deref() {
+        if is_explicit_onboard_clear_input(api_key_env) {
+            return Ok(String::new());
+        }
+        let trimmed = api_key_env.trim();
+        if trimmed.is_empty() {
+            None
+        } else {
+            Some(validate_selected_provider_credential_env(config, trimmed)?)
+        }
+    } else {
+        None
+    };
+
+    if options.non_interactive {
+        return Ok(explicit_selection.unwrap_or(default_api_key_env));
+    }
+    let initial = explicit_selection
+        .as_deref()
+        .unwrap_or(default_api_key_env.as_str());
+    let example_env_name =
+        provider_credential_policy::provider_credential_env_hint(&config.provider)
+            .unwrap_or_else(|| "PROVIDER_API_KEY".to_owned());
+    loop {
+        print_lines(
+            ui,
+            render_api_key_env_selection_screen_lines_with_style(
+                config,
+                default_api_key_env.as_str(),
+                initial,
+                guided_prompt_path,
+                context.render_width,
+                true,
+            ),
+        )?;
+        let value = ui.prompt_with_default("Credential env var name", initial)?;
+        if is_explicit_onboard_clear_input(&value) {
+            return Ok(String::new());
+        }
+        let trimmed = value.trim();
+        if trimmed.is_empty() {
+            return Ok(String::new());
+        }
+        match validate_selected_provider_credential_env(config, trimmed) {
+            Ok(validated) => return Ok(validated),
+            Err(error) => {
+                print_message(ui, error)?;
+                print_message(
+                    ui,
+                    format!(
+                        "enter the environment variable name only, for example {example_env_name}, or type :clear to remove the env binding"
+                    ),
+                )?;
+            }
+        }
+    }
+}
+
+pub(super) fn apply_selected_api_key_env(
+    provider: &mut mvp::config::ProviderConfig,
+    selected_api_key_env: String,
+) {
+    let selected_api_key_env = selected_api_key_env.trim();
+    if selected_api_key_env.is_empty() {
+        provider.clear_api_key_env_binding();
+        provider.clear_oauth_access_token_env_binding();
+        return;
+    }
+
+    provider.api_key = None;
+    provider.oauth_access_token = None;
+    match provider_credential_policy::selected_provider_credential_env_field(
+        provider,
+        selected_api_key_env,
+    ) {
+        provider_credential_policy::ProviderCredentialEnvField::ApiKey => {
+            provider.clear_oauth_access_token_env_binding();
+            provider.set_api_key_env_binding(Some(selected_api_key_env.to_owned()));
+        }
+        provider_credential_policy::ProviderCredentialEnvField::OAuthAccessToken => {
+            provider.clear_api_key_env_binding();
+            provider.set_oauth_access_token_env_binding(Some(selected_api_key_env.to_owned()));
+        }
+    }
+}
+
+pub(super) fn onboard_credential_env_name_is_safe(raw: &str) -> bool {
+    let trimmed = raw.trim();
+    if trimmed.is_empty() {
+        return false;
+    }
+
+    let mut config = mvp::config::LoongConfig::default();
+    config.provider.api_key = Some(SecretRef::Env {
+        env: trimmed.to_owned(),
+    });
+    config.provider.api_key_env = None;
+
+    config.validate().is_ok()
+}
